feat(assistant): allow agnostic provider
