+++ layer/modules/assistant/index.ts
@@ -1,7 +1,23 @@
-import { addComponent, addImports, addServerHandler, createResolver, defineNuxtModule, logger } from '@nuxt/kit'
+import { addComponent, addImports, addServerHandler, addServerImports, createResolver, defineNuxtModule, logger } from '@nuxt/kit'
 import { defu } from 'defu'
 
 export interface AssistantModuleOptions {
+  /**
+   * Enable the assistant.
+   *
+   * When left undefined, the assistant is enabled if `AI_GATEWAY_API_KEY` or
+   * `VERCEL_OIDC_TOKEN` is available at build time.
+   *
+   * Set it to `true` to use a custom AI SDK provider: Docus then skips
+   * registering its own endpoint, and you provide a route at `apiPath` built
+   * with `assistantSearchHandler`.
+   *
+   * Set it to `false` to disable the assistant even when AI Gateway
+   * credentials are available.
+   *
+   * @default undefined
+   */
+  enabled?: boolean
   /**
    * API endpoint path for the assistant
    * @default '/__docus__/assistant'
@@ -21,9 +37,15 @@ export interface AssistantModuleOptions {
   model?: string
 }
 
+/** Subset of the Nitro instance used to hand the endpoint over to a user route. */
+interface NitroBuildContext {
+  scannedHandlers: Array<{ route?: string }>
+  options: { handlers: Array<{ route?: string, handler?: string }> }
+}
+
 const log = logger.withTag('docus')
 
-const defaults: Required<AssistantModuleOptions> = {
+const defaults: Required<Omit<AssistantModuleOptions, 'enabled'>> = {
   apiPath: '/__docus__/assistant',
   mcpServer: '/mcp',
   model: 'google/gemini-3-flash',
@@ -45,10 +67,17 @@ export default defineNuxtModule<AssistantModuleOptions>({
       process.env.AI_GATEWAY_API_KEY || process.env.VERCEL_OIDC_TOKEN
     )
 
+    // An explicit `enabled` value always wins, so a custom AI SDK provider can
+    // be used without AI Gateway credentials.
+    const isEnabled = options.enabled ?? hasAiGatewayAuth
+    // Docus only owns the endpoint when it can authenticate to the AI Gateway.
+    // Otherwise the user brings their own route built with `assistantSearchHandler`.
+    const hasDefaultHandler = isEnabled && hasAiGatewayAuth
+
     const { resolve } = createResolver(import.meta.url)
 
     nuxt.options.runtimeConfig.public.assistant = {
-      enabled: hasAiGatewayAuth,
+      enabled: isEnabled,
       apiPath: options.apiPath,
     }
 
@@ -68,7 +97,7 @@ export default defineNuxtModule<AssistantModuleOptions>({
     components.forEach(name =>
       addComponent({
         name,
-        filePath: hasAiGatewayAuth
+        filePath: isEnabled
           ? resolve(`./runtime/components/${name}.vue`)
           : resolve('./runtime/components/AssistantChatDisabled.vue'),
       }),
@@ -79,22 +108,66 @@ export default defineNuxtModule<AssistantModuleOptions>({
       filePath: resolve('./runtime/components/AssistantComark'),
     })
 
-    if (!hasAiGatewayAuth) {
-      nuxt.hook('modules:done', () => {
-        log.warn('AI assistant disabled: neither `AI_GATEWAY_API_KEY` nor `VERCEL_OIDC_TOKEN` found')
-      })
-      return
-    }
-
     nuxt.options.runtimeConfig.assistant = {
       mcpServer: options.mcpServer,
       model: options.model,
     }
 
+    // Exposed even when disabled so overriding the endpoint stays type-safe.
+    addServerImports([
+      {
+        name: 'assistantSearchHandler',
+        from: resolve('./runtime/server/utils/assistant'),
+      },
+      {
+        name: 'getAssistantSystemPrompt',
+        from: resolve('./runtime/server/utils/assistant'),
+      },
+    ])
+
+    if (!isEnabled) {
+      if (options.enabled === undefined) {
+        nuxt.hook('modules:done', () => {
+          log.warn('AI assistant disabled: neither `AI_GATEWAY_API_KEY` nor `VERCEL_OIDC_TOKEN` found')
+        })
+      }
+      return
+    }
+
+    if (!hasDefaultHandler) {
+      nuxt.hook('modules:done', () => {
+        log.info(`AI assistant enabled without AI Gateway credentials: provide a server route at \`${options.apiPath}\` using \`assistantSearchHandler\``)
+      })
+      return
+    }
+
     const routePath = options.apiPath!.replace(/^\//, '')
-    addServerHandler({
-      route: `/${routePath}`,
-      handler: resolve('./runtime/server/api/search'),
+    const route = `/${routePath}`
+    const handler = resolve('./runtime/server/api/search')
+
+    addServerHandler({ route, handler })
+
+    // A server route defined by the user at the same path would otherwise be
+    // silently shadowed by the handler above, so drop ours when it exists.
+    // `nitro:build:before` is declared by `@nuxt/nitro-server`, which Docus does
+    // not depend on directly, hence the local typing.
+    const hookNitroBuild = nuxt.hook as unknown as (
+      name: 'nitro:build:before',
+      callback: (nitro: NitroBuildContext) => void,
+    ) => void
+
+    hookNitroBuild('nitro:build:before', (nitro) => {
+      if (!nitro.scannedHandlers.some(scanned => scanned.route === route)) {
+        return
+      }
+
+      const index = nitro.options.handlers.findIndex(h => h.route === route && h.handler === handler)
+      if (index === -1) {
+        return
+      }
+
+      nitro.options.handlers.splice(index, 1)
+      log.info(`AI assistant using your \`${route}\` server route instead of the built-in endpoint`)
     })
   },
 })
